+++ fivox/genericLoader.h
@@ -47,7 +47,6 @@ private:
     Vector2f _getTimeRange() const final;
     ssize_t _load( size_t chunkIndex, size_t numChunks ) final;
     SourceType _getType() const final { return SourceType::frame; }
-    bool _hasEnded() const final { return true; }
     size_t _getNumChunks() const final { return 1; }
     //@}