+++ fivox/spikeLoader.cpp
@@ -24,12 +24,8 @@
 
 #include <brain/brain.h>
 #include <brion/brion.h>
-#include <brain/spikeReportReader.h>
-#include <brain/spikes.h>
+
 #include <lunchbox/os.h>
-#include <lunchbox/lock.h>
-#include <lunchbox/scopedMutex.h>
-#include <lunchbox/memoryMap.h>
 
 using boost::lexical_cast;
 
@@ -40,9 +36,7 @@ class SpikeLoader::Impl
 public:
     Impl( EventSource& output, const URIHandler& params )
         : _output( output )
-        , _duration( params.getDuration( ))
         , _spikesStart( 0.f )
-        , _spikesEnd( 0.f )
     {
         const brion::GIDSet& gids = params.getGIDs();
 
@@ -60,62 +54,28 @@ public:
         }
 
         const std::string& spikePath = params.getSpikes();
-        _loadSpikes( spikePath.empty() ? params.getConfig().getSpikeSource()
-                                       : URI( spikePath ));
-
+        _report.reset(
+            new brain::SpikeReportReader( spikePath.empty()
+                                          ? params.getConfig().getSpikeSource()
+                                          : URI( spikePath ),
+                                          gids ));
+        _spikesEnd = _report->getEndTime();
     }
 
     void updateTimeRange()
     {
-        // Streaming in progress; Only report fully finished frames.
-        if( _spikesReader && !_spikesReader->hasEnded( ))
-        {
-            lunchbox::ScopedWrite mutex( _getSpikesLock );
-            const brain::Spikes& spikes = _spikesReader->getSpikes();
-            if( !spikes.empty( ))
-                // don't update _spikesStart to calculate absolute frame numbers
-                // see https://bbpcode.epfl.ch/code/#/c/19337
-                _spikesEnd = spikes.getEndTime();
-        }
-    }
-
-    void _loadSpikes( const URI& spikes )
-    {
-        if( _loadBinarySpikes( spikes.getPath( )))
+        if( !_report || _report->hasEnded( ))
             return;
 
-        LBINFO << "No valid binary .spikes file found, loading from .dat..."
-               << std::endl;
-        _spikesReader.reset( new brain::SpikeReportReader( spikes ));
-        _spikesStart = _spikesReader->isStream() ? 0.f :
-                                                _spikesReader->getStartTime();
-        _spikesEnd = _spikesReader->isStream() ? 0.f :
-                                                _spikesReader->getEndTime();
-    }
-
-    bool _loadBinarySpikes( const std::string& spikes )
-    {
-        std::unique_ptr< lunchbox::MemoryMap > spikesFile
-                ( new lunchbox::MemoryMap( spikes ));
-        const size_t size = spikesFile->getSize();
-        if(( size % sizeof( uint32_t )) != 0 )
-            return false;
-        const size_t nElems = size / sizeof( uint32_t );
-        const uint32_t* iData = spikesFile->getAddress< uint32_t >();
-        size_t index = 0;
-
-        const uint32_t magic = 0xf0a;
-        const uint32_t version = 1;
-        if( index >= nElems || iData[ index++ ] != magic )
-            return false;
-        if( index >= nElems || iData[ index++ ] != version )
-            return false;
-
-        _spikesFile = std::move( spikesFile );
-        const float* fData = _spikesFile->getAddress< float >();
-        _spikesStart = fData[index]; // first spike timestamp after header
-        _spikesEnd = fData[nElems - 2]; // last spike timestamp
-        return true;
+        const auto endtime = _report->getEndTime( );
+        // This forces the collection of the latest spikes in the stream
+        // case, thus updating the end time.
+        _report->getSpikes(
+            std::nextafter( endtime, - std::numeric_limits<float>::max( )),
+            endtime );
+        // don't update _spikesStart to calculate absolute frame numbers
+        // see https://bbpcode.epfl.ch/code/#/c/19337
+        _spikesEnd = _report->getEndTime();
     }
 
     ssize_t load()
@@ -124,9 +84,8 @@ public:
         lunchbox::setZero( _spikesPerNeuron.data(),
                            _spikesPerNeuron.size() * sizeof(size_t));
 
-        const float end = start + _duration;
-        const size_t numSpikes = _spikesFile ? _loadSpikesFast( start, end )
-                                             : _loadSpikesSlow( start, end );
+        const float end = start + _output.getDuration();
+        const size_t numSpikes = _loadSpikes( start, end );
 
         for( size_t i = 0; i < _spikesPerNeuron.size(); ++i )
             _output[i] = _spikesPerNeuron[i];
@@ -134,56 +93,12 @@ public:
         return numSpikes;
     }
 
-    // OPT: directly iterate on binary spike file; saves loading all spikes
-    // a priori and slow access in multimap (brion::Spikes)
-    size_t _loadSpikesFast( const float start, const float end )
-    {
-        const size_t size = _spikesFile->getSize();
-        const size_t nElems = size / sizeof( uint32_t );
-        const uint32_t* iData = _spikesFile->getAddress< uint32_t >();
-        const float* fData = _spikesFile->getAddress< float >();
 
-        if( start < _previousStart.first )
-            _previousStart.second = 2; // start over, first spike after header
-
-        size_t numSpikes = 0;
-        bool once = true;
-        for( size_t i = _previousStart.second; i+1 < nElems; i += 2 )
-        {
-            const float time = fData[ i ];
-            if( time < start )
-                continue;
-            if( once )
-            {
-                // remember this time for next start that might come after
-                _previousStart = std::make_pair( start, i );
-                once = false;
-            }
-
-            if( time >= end )
-                break;
-
-            const uint32_t gid = iData[ i + 1 ];
-            if( gid >= _gidIndex.size( ))
-                continue;
-
-            ++_spikesPerNeuron[_gidIndex[gid]];
-            ++numSpikes;
-        }
-        return numSpikes;
-    }
-
-    // for ~5 mio spikes, this is ~200ms slower than _loadSpikesFast
-    size_t _loadSpikesSlow( const float start, const float end )
+    size_t _loadSpikes( const float start, const float end )
     {
         size_t numSpikes = 0;
-        lunchbox::ScopedWrite mutex( _getSpikesLock );
-        const brain::Spikes& spikes = _spikesReader->getSpikes( start, end );
-        for( const brion::Spike& spike : spikes )
+        for( const auto& spike : _report->getSpikes( start, end ))
         {
-            if( spike.second >= _gidIndex.size( ))
-                continue;
-
             ++_spikesPerNeuron[_gidIndex[spike.second]];
             ++numSpikes;
         }
@@ -192,7 +107,6 @@ public:
     }
 
     EventSource& _output;
-    const float _duration;
     float _spikesStart;
     float _spikesEnd;
 
@@ -206,13 +120,7 @@ public:
     // (container.size() is number of GIDs)
     brion::size_ts _spikesPerNeuron;
 
-    // for _loadSpikesFast
-    std::unique_ptr< lunchbox::MemoryMap > _spikesFile;
-    std::pair< float, size_t > _previousStart;
-
-    // for _loadSpikesSlow
-    std::unique_ptr< brain::SpikeReportReader > _spikesReader;
-    mutable lunchbox::Lock _getSpikesLock;
+    std::unique_ptr< brain::SpikeReportReader > _report;
 };
 
 SpikeLoader::SpikeLoader( const URIHandler& params )
@@ -229,25 +137,13 @@ SpikeLoader::~SpikeLoader()
 Vector2f SpikeLoader::_getTimeRange() const
 {
     _impl->updateTimeRange();
-
-    // The duration of the frame needs to be considered,
-    // in order to not go over the available range.
-    const float spikesEnd = _impl->_spikesEnd - _impl->_duration;
-    if( spikesEnd < _impl->_spikesStart )
-        return Vector2f( 0.f, 0.f );
-
-    return Vector2f( _impl->_spikesStart, spikesEnd );
+    return Vector2f( _impl->_spikesStart, _impl->_spikesEnd );
 }
 
 ssize_t SpikeLoader::_load( const size_t /*chunkIndex*/,
                             const size_t /*numChunks*/ )
 {
     return _impl->load();
-}
-
-bool SpikeLoader::_hasEnded() const
-{
-    return _impl->_spikesReader ? _impl->_spikesReader->hasEnded() : true;
 }
 
 }